Add files via upload
Map projection software

These are the source files associated with the Instructable here

https://www.instructables.com/WorldClock

Or search www.instructables.com for "World Clock"

The latest versions of the files will be in the Instructable. The versions here may have been superseded.

However, Instructables does not allow uploading of EXE files, hence these files are here

You should refer to the Instructable to find how to use these files.